fix(email): send attendee ticket emails on dashboard 'Mark as Paid'
The dashboard 'Mark as Paid' button uses a server action defined inside
the order detail page, not the /api/orders/[id]/mark-paid API route. PR
#358 wired sendAttendeeTicketEmailsForOrder into the API route but
missed the server action, so manual orders only emailed the buyer.

Mirror the API-route behaviour in the server action: the buyer keeps
getting the full receipt + all tickets, and every non-buyer attendee
now also receives a ticket-only PDF email.

Also drop the /api/orders/[id]/mark-paid route — it has no callers.
